pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sort-it-now"
version = "0.1.0"
description = "Tray-resident real-time file organizer assistant"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "watchdog>=3.0.0,<5.0.0",
    "pystray>=0.19.0,<1.0.0",
    "Pillow>=10.0.0,<12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<9.0.0",
    "pytest-cov>=4.0.0,<6.0.0",
    "pyinstaller>=6.0.0,<7.0.0",
]

[project.scripts]
sort-it-now = "sort_it_now.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools.packages.find]
where = ["src"]
